Move ui tests under commonTest module
This is the recommended way to write common UI tests:
https://www.jetbrains.com/help/kotlin-multiplatform-dev/compose-test.html#write-and-run-common-tests

For now, we can run all the tests on Desktop and iOS, but not on Android.
There’s currently no separation between UI and unit tests since they’re in the same package.
We could exclude the folder from unit tests, but I won’t do that — from now on, it’s fine to run all tests together.

This will be helpful for CI, and I’ll check if running tests in a Desktop environment is faster.
Also, avoiding the Android emulator should help with CI storage and overall duration.